refactor(ai-agent): move BuildExtra types to aionui-api-types
